Factored out get_mutated_variables_in_fn to own helper module.
